+++ elfo-core/src/logging/config.rs
@@ -7,7 +7,7 @@ use tracing::level_filters::LevelFilter;
 pub(crate) struct LoggingConfig {
     #[serde(deserialize_with = "deserialize_level_filter")]
     pub(crate) max_level: LevelFilter,
-    pub(crate) max_rate: u64,
+    pub(crate) max_rate_per_level: u64,
     pub(crate) targets: FxHashMap<String, LoggingTargetConfig>,
 }
 
@@ -49,7 +49,7 @@ impl Default for LoggingConfig {
     fn default() -> Self {
         Self {
             max_level: LevelFilter::INFO,
-            max_rate: 1000,
+            max_rate_per_level: 1000,
             targets: FxHashMap::default(),
         }
     }